fixes rubocop and erblint
