+++ app/views/alerts/create.turbo_stream.erb
@@ -1,2 +1,1 @@
 <%= turbo_stream.replace "new_alert_form", partial: "alerts/form", locals: { alert: Alert.new } %>
-
